requirements.txt
opensimplex==0.4.3
tracery==0.1.1
perlin-noise==1.12
pixelsort==1.0.1
Pillow==9.2.0
deap==1.3.3